refatora: as lógicas associadas as entidades Usuario, Local e Tag para que se adequem ao design

+++ backend/urbanmap-api/src/main/java/br/com/senac/urbanmap/controllers/LocalController.java
@@ -1,9 +1,7 @@
 package br.com.senac.urbanmap.controllers;
 
-import br.com.senac.urbanmap.controllers.dtos.LocalAlteracaoDTO;
-import br.com.senac.urbanmap.controllers.dtos.LocalCadastroDTO;
-import br.com.senac.urbanmap.controllers.dtos.LocalPadraoDTO;
-import br.com.senac.urbanmap.controllers.dtos.LocalListaReduzidaDTO;
+import br.com.senac.urbanmap.controllers.dtos.*;
+import br.com.senac.urbanmap.controllers.dtos.local.*;
 import br.com.senac.urbanmap.entities.local.Local;
 import br.com.senac.urbanmap.services.LocalService;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -29,46 +27,76 @@ public class LocalController {
         this.localService = localService;
     }
 
-    @GetMapping("/todos")
-    public ResponseEntity<List<LocalListaReduzidaDTO>> buscarTodos() {
-        return ResponseEntity.ok(LocalListaReduzidaDTO.converterParaListaDTO(this.localService.buscarTodos()));
+    // USUARIO COMUM
+
+    @GetMapping("/lista")
+    public ResponseEntity<?> buscarTodos(@RequestParam("usuarioID") Long usuarioID) {
+        return ResponseEntity.ok(this.localService.obterPainelUsuario(usuarioID));
     }
 
-    // nome ou tags
     @GetMapping("/buscar")
-    public ResponseEntity<List<LocalListaReduzidaDTO>> buscarPorParametro(
+    public ResponseEntity<List<LocalResumidoUsuarioDTO>> buscarPorParametro(
             @RequestParam(value = "nome", required = false) String nome,
             @RequestParam(value = "idTags", required = false) Set<Long> idTags) {
-        return ResponseEntity.ok(LocalListaReduzidaDTO.converterParaListaDTO(this.localService.buscarPorParametros(nome, idTags)));
+
+        List<LocalResumidoUsuarioDTO> locais = this.localService.buscarPorParametrosUsuario(nome, idTags);
+        return ResponseEntity.ok(locais);
+    }
+
+    @GetMapping("/usuario/{idUsuario}/curtidos")
+    public ResponseEntity<List<LocalResumidoUsuarioDTO>> listarCurtidos(@PathVariable Long idUsuario) {
+        return ResponseEntity.ok(this.localService.obterLocaisCurtidos(idUsuario));
+    }
+
+    @GetMapping("/usuario/{idUsuario}/salvos")
+    public ResponseEntity<List<LocalResumidoUsuarioDTO>> listarSalvos(@PathVariable Long idUsuario) {
+        return ResponseEntity.ok(this.localService.obterLocaisSalvos(idUsuario));
     }
 
     @GetMapping("/buscar/{idLocal}")
-    public ResponseEntity<LocalPadraoDTO> buscarLocal(@PathVariable("idLocal") Long idLocal) {
-        return ResponseEntity.ok(LocalPadraoDTO.converterParaDto(this.localService.buscarLocal(idLocal)));
+    public ResponseEntity<LocalCompletoDTO> buscarLocal(@PathVariable("idLocal") Long idLocal) {
+        return ResponseEntity.ok(LocalCompletoDTO.converterParaDto(this.localService.buscarLocal(idLocal)));
     }
 
-    @PostMapping(value = "/cadastrar", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<LocalPadraoDTO> cadastrar(
+    // ADMIN
+
+    @GetMapping("/admin/painel")
+    public ResponseEntity<LocalPainelAdminDTO> obterPainel() {
+        return ResponseEntity.ok(this.localService.obterPainelAdmin());
+    }
+
+    @GetMapping("/admin/estatistica/geral")
+    public ResponseEntity<EstatisticaDTO> obterEstatisticaGeral() {
+        return ResponseEntity.ok(this.localService.estatisticaGeral());
+    }
+
+    @PostMapping(value = "/admin/cadastrar", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    public ResponseEntity<LocalCompletoDTO> cadastrar(
             @Parameter(content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE))
             @RequestPart("local") @Valid LocalCadastroDTO dto,
             @RequestPart(value = "arquivos", required = false) List<MultipartFile> arquivos) {
         Local local = this.localService.cadastrar(dto, arquivos);
-        return ResponseEntity.created(URI.create("/locais/" + local.getId())).body(LocalPadraoDTO.converterParaDto(local));
+        return ResponseEntity.created(URI.create("/locais/" + local.getId())).body(LocalCompletoDTO.converterParaDto(local));
     }
 
-    @PutMapping(value = "/alterar", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    @PutMapping(value = "/admin/alterar", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<?> alterar(
             @Parameter(content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE))
             @RequestPart("local") @Valid LocalAlteracaoDTO dto,
             @RequestPart(value = "arquivos", required = false) List<MultipartFile> arquivos) {
+        if (!this.localService.localExiste(dto.id()))
+            return ResponseEntity.notFound().build();
+
         this.localService.alterar(dto, arquivos);
         return ResponseEntity.ok().build();
     }
 
-    @DeleteMapping("/deletar/{idLocal}")
+    @DeleteMapping("/admin/deletar/{idLocal}")
     public ResponseEntity<?> deletarLocal(@PathVariable("idLocal") Long idLocal) {
+        if (!this.localService.localExiste(idLocal))
+            return ResponseEntity.notFound().build();
+
         this.localService.remover(idLocal);
         return ResponseEntity.noContent().build();
     }
-
 }
